@@ -36,7 +36,7 @@ class _NSAMainFormState extends State<NSAMainForm> {
     double deviceHeight = MediaQuery.of(context).size.height;
     return WillPopScope(
       onWillPop: () {
-        return Future.value(false);
+        return Future.value(true);
       },
       child: Scaffold(
         // resizeToAvoidBottomInset: false,
