@@ -652,12 +652,14 @@ class _PrivilegeMainV3 extends State<PrivilegeMainV3>
                   color: Color(0x99FFFFFF),
                   borderRadius: BorderRadius.circular(15),
                 ),
-                child: BackdropFilter(
-                  filter: ImageFilter.blur(
-                    sigmaX: 5.0,
-                    sigmaY: 5.0,
-                  ),
-                  child: Column(
+                child: 
+                // BackdropFilter(
+                //   filter: ImageFilter.blur(
+                //     sigmaX: 5.0,
+                //     sigmaY: 5.0,
+                //   ),
+                //   child: 
+                  Column(
                     mainAxisAlignment: MainAxisAlignment.spaceBetween,
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: [
@@ -677,7 +679,7 @@ class _PrivilegeMainV3 extends State<PrivilegeMainV3>
                       ),
                     ],
                   ),
-                ),
+                // ),
               ),
             ),
           )
